tools/generator-terraform: splitting the existing tests out per transform type to make this more readable
